Implementando comunicaçao no banco de dados

+++ WebProdutos/Migrations/ProdutoContextModelSnapshot.cs
@@ -37,7 +37,7 @@ namespace WebProdutos.Migrations
                         .IsRequired()
                         .HasColumnType("text");
 
-                    b.Property<decimal>("valor")
+                    b.Property<float>("valor")
                         .HasColumnType("integer");
 
                     b.HasKey("Id");
